Properly document methods and change date_utils year to use u16

+++ src/date_utils.rs
@@ -1,4 +1,5 @@
-pub fn is_leap_year(year: u32) -> bool {
+/// Takes a year as a u16 and returns whether it is a leap year.
+pub fn is_leap_year(year: u16) -> bool {
     (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0))
 }
 
@@ -6,8 +7,8 @@ pub fn is_leap_year(year: u32) -> bool {
 mod tests {
     #[test]
     fn test_leap_years() {
-        let valid: [u32; 3] = [2000, 2012, 2016];
-        let invalid: [u32; 3] = [2100, 2018, 2013];
+        let valid: [u16; 3] = [2000, 2012, 2016];
+        let invalid: [u16; 3] = [2100, 2018, 2013];
         for v in valid.iter() {
             assert!(super::is_leap_year(*v));
         }
